+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Issue [#546](https://github.com/m-marini/wheellyj/issues/546): Mapping state
-- Issue [#552](https://github.com/m-marini/wheellyj/issues/546): Add Astar-RRT search state
+- Issue [#552](https://github.com/m-marini/wheellyj/issues/542): Add Astar-RRT search state
+- Issue [#554](https://github.com/m-marini/wheellyj/issues/544): Add label stuck state
 
 ### Changed